bgfx: fix disablewhen comparisons against non-zero values never matching
suppressor_reader stored comparison values as raw int bit patterns,
but bgfx_suppressor::suppress() memcmp()s them against float slider
values (bgfx_slider::value() returns float for every slider type).
int 1 is 0x00000001 while float 1.0f is 0x3F800000, so any equality
test against a non-zero value - "color_saturation == 1.0",
"chroma_mode == 3", "power == [1,1,1]" - could never match and the
pass ran every frame; only ==0 comparisons worked, via the shared
all-zero bit pattern. Store the values as floats instead, and accept
non-integer numbers in array values (previously IsInt-only).

No chain uses the notequal condition, so this only makes the
intended identity-skips actually skip.

+++ src/osd/modules/render/bgfx/suppressorreader.h
@@ -25,7 +25,7 @@ public:
 	static bgfx_suppressor* read_from_value(const Value& value, const std::string &prefix, std::map<std::string, bgfx_slider*>& sliders);
 
 private:
-	static bool get_values(const Value& value, std::string prefix, std::string name, int* values, const int count);
+	static bool get_values(const Value& value, std::string prefix, std::string name, float* values, const int count);
 	static bool validate_parameters(const Value& value, const std::string &prefix);
 
 	static const int CONDITION_COUNT = 2;
